ExponentialFamilyManifolds

ExponentialFamilyManifolds.jl provides implementations of manifolds for the natural parameters of exponential family distributions, using Manifolds.jl. These manifolds are compatible with ManifoldsBase.jl, enabling optimization of the natural parameters of exponential family distributions using Manopt.jl.

ExponentialFamilyManifolds.Negated

ExponentialFamilyManifolds.get_natural_manifold

ExponentialFamilyManifolds.get_natural_manifold_base

ExponentialFamilyManifolds.partition_point

ExponentialFamilyManifolds.Negated — Type

Negated(m)

Lazily negates the matrix m, without creating a new matrix. Works by redefining the getindex.

ExponentialFamilyManifolds.get_natural_manifold — Method

get_natural_manifold(::Type{T}, conditioner = nothing)

The function returns a corresponding manifold for the natural parameters of distribution of type T. Optionally accepts the conditioner, which is set to nothing by default.

ExponentialFamilyManifolds.get_natural_manifold_base — Function

get_natural_manifold_base(::Type{T}, conditioner = nothing)

Returns base manifold for the distribution of type T.

ExponentialFamilyManifolds.partition_point — Method

partition_point(M::NaturalParametersManifold, point)
partition_point(::Type{T}, conditioner, point)

Converts point to a compatible representation for the natural manifold M of type T.