ExponentialFamilyManifolds

ExponentialFamilyManifolds.jl provides implementations of manifolds for the natural parameters of exponential family distributions, using Manifolds.jl. These manifolds are compatible with ManifoldsBase.jl, enabling optimization of the natural parameters of exponential family distributions using Manopt.jl.

ExponentialFamilyManifolds.NaturalParametersManifold

ExponentialFamilyManifolds.Negated

ExponentialFamilyManifolds.ShiftedNegativeNumbers

ExponentialFamilyManifolds.ShiftedPositiveNumbers

ExponentialFamilyManifolds.SymmetricNegativeDefinite

ExponentialFamilyManifolds.get_natural_manifold

ExponentialFamilyManifolds.get_natural_manifold_base

ExponentialFamilyManifolds.partition_point

Manifolds

Distribution specific manifolds

ExponentialFamilyManifolds.get_natural_manifold — Function

get_natural_manifold(::Type{T}, dims, conditioner = nothing)

The function returns a corresponding manifold for the natural parameters of distribution of type T. Optionally accepts the conditioner, which is set to nothing by default. Use empty tuple () for univariate distributions.

julia> using ExponentialFamily, ExponentialFamilyManifolds

julia> ExponentialFamilyManifolds.get_natural_manifold(Beta, ()) isa ExponentialFamilyManifolds.NaturalParametersManifold
true

julia> ExponentialFamilyManifolds.get_natural_manifold(MvNormalMeanCovariance, (3, )) isa ExponentialFamilyManifolds.NaturalParametersManifold
true

ExponentialFamilyManifolds.NaturalParametersManifold — Type

NaturalParametersManifold(::Type{T}, dims, base, conditioner)

The manifold for the natural parameters of the distribution of type T with dimensions dims. An internal structure, use get_natural_manifold to create an instance of a manifold for the natural parameters of distribution of type T.

ExponentialFamilyManifolds.get_natural_manifold_base — Function

get_natural_manifold_base(::Type{T}, dims, conditioner = nothing)

Returns base manifold for the distribution of type T of dimension dims. Optionally accepts the conditioner, which is set to nothing by default.

ExponentialFamilyManifolds.partition_point — Function

partition_point(M::NaturalParametersManifold, dims, p)
partition_point(::Type{T}, dims, point, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold M of type T.

partition_point(::Type{Bernoulli}, ::Tuple{}, p, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold of type Bernoulli.

partition_point(::Type{Beta}, ::Tuple{}, p, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold of type Beta.

partition_point(::Type{Chisq}, ::Tuple{}, p, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold of type Chisq.

partition_point(::Type{Dirichlet}, dims::Tuple{Int}, p, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold of type Dirichlet.

partition_point(::Type{Exponential}, ::Tuple{}, p, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold of type Exponential.

partition_point(::Type{Gamma}, ::Tuple{}, p, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold of type Gamma.

partition_point(::Type{Geometric}, ::Tuple{}, p, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold of type Geometric.

partition_point(::Type{Laplace}, ::Tuple{}, p, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold of type Laplace.

partition_point(::Type{LogNormal}, ::Tuple{}, p, conditioner=nothing)

Converts the point to a compatible representation for the natural manifold of type LogNormal.

partition_point(::Type{NormalMeanVariance}, ::Tuple{}, p, conditioner = nothing)

Converts the point to a compatible representation for the natural manifold of type NormalMeanVariance.

partition_point(::Type{MvNormalMeanCovariance}, dims::Tuple{Int}, p, conditioner = nothing)

Converts the point to a compatible representation for the natural manifold of type MvNormalMeanCovariance.

Custom generic manifolds

ExponentialFamilyManifolds.jl provides some extra manifolds, which are not included in the Manifolds.jl

ExponentialFamilyManifolds.ShiftedPositiveNumbers — Type

ShiftedPositiveNumbers(shift)

A manifold representing the positive numbers shifted by shift. The points on this manifold are 1-dimensional vectors with a single element.

ExponentialFamilyManifolds.ShiftedNegativeNumbers — Type

ShiftedNegativeNumbers(shift)

A manifold representing the negative numbers shifted by shift. The points on this manifold are 1-dimensional vectors with a single element.

ExponentialFamilyManifolds.SymmetricNegativeDefinite — Type

SymmetricNegativeDefinite(k)

This manifold represents the set of negative definite matrices of size k × k. Similar to SymmetricPositiveDefinite from Manifolds.jl with the exception that the matrices are negative definite.

Helpers

ExponentialFamilyManifolds.Negated — Type

Negated(m)

Lazily negates the matrix m, without creating a new matrix. Works by redefining the getindex.